refactor: detail viewcontroller imageheight를 CGFloat으로 수정

+++ Expo1900/Expo1900/Controller/ExpositionItemDetailViewController.swift
@@ -25,7 +25,7 @@ final class ExpositionItemDetailViewController: UIViewController {
     }
     
     private func configureUI() {
-        let imageHeight = 150
+        let imageHeight = CGFloat(150)
         detailImageView.image = UIImage(named: expositionItem?.imageName ?? "")?.resized(height: imageHeight)
         
         detailDescriptionLabel.text = expositionItem?.description
